Move DataState to new swap package
