+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swap_new/approve/SwapApproveService.kt
@@ -2,7 +2,7 @@ package io.horizontalsystems.bankwallet.modules.swap_new.approve
 
 import io.horizontalsystems.bankwallet.core.Clearable
 import io.horizontalsystems.bankwallet.core.ethereum.EthereumTransactionService
-import io.horizontalsystems.bankwallet.modules.swap.DataState
+import io.horizontalsystems.bankwallet.entities.DataState
 import io.horizontalsystems.erc20kit.core.Erc20Kit
 import io.horizontalsystems.ethereumkit.core.EthereumKit
 import io.horizontalsystems.ethereumkit.models.Address
